arangopipe/tests/CItests/requirements.txt
python-arango
arangopipe
pandas
PyYAML==5.1.1
sklearn2

